ai-task-orchestrator/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
